FIX: dto types validations and seperate files
