+++ src/models/filter-by.interface.ts
