+++ src/news/dtos/filter-by.dto.ts
@@ -1,13 +1,14 @@
-import { IsString, ValidateNested } from "class-validator"
+import { IsAlpha, IsNotEmpty, IsString, ValidateNested } from "class-validator"
 import { DatesDto } from "./dates-dto"
 import { Schema } from '@nestjs/mongoose'
 import { Type } from 'class-transformer'
 
 
 
-@Schema()
 export class FilterByDto {
+    @IsNotEmpty()
     @IsString()
+    @IsAlpha()
     type: string
 
     @IsString()
