+++ src/news/dtos/search-articles.dto.ts
@@ -1,8 +1,8 @@
-import { IsString, IsNumber, ValidateNested, IsNotEmpty } from "class-validator"
+import { IsString, IsNumber, ValidateNested, IsNotEmpty, IsPositive, IsInt, IsDivisibleBy} from "class-validator"
 import { Type } from 'class-transformer'
 import { FilterByDto } from "./filter-by.dto"
 
-export class CreateArticleDto {
+export class SearchArticlesDto {
 
     @ValidateNested()
     @IsNotEmpty()
@@ -13,6 +13,10 @@ export class CreateArticleDto {
     searchQuery: string
 
     @IsNumber()
+    @IsInt()
+    @IsPositive()
+    @IsDivisibleBy(1)
+
     page: number
 
 }
